Separate activation tests from registration tests. (#25)
